arrumando o codigo

+++ src/main/java/br/com/fiap/projetoestagio/repository/CadastroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.projetoestagio.models.Cadastro;
 
-public interface CadastroRepository extends JpaRepository<Cadastro, Long>{
+public interface CadastroRepository extends JpaRepository<Cadastro, Integer>{
     
 }
